[WFS provider] Split qgsbackgroundcachedfeatureiterator.h/cpp into as many files as classes
No code change. Just better reflect the class organization
