+++ src/providers/wfs/qgsbackgroundcachedfeatureiterator.cpp
@@ -1,6 +1,6 @@
 /***************************************************************************
-    qgsbackgroundcachedfeatureiterator.h
-    ---------------------
+    qgsbackgroundcachedfeatureiterator.cpp
+    --------------------------------------
     begin                : October 2019
     copyright            : (C) 2016-2019 by Even Rouault
     email                : even.rouault at spatialys.com
@@ -15,246 +15,30 @@
 
 #include "qgsbackgroundcachedfeatureiterator.h"
 #include "moc_qgsbackgroundcachedfeatureiterator.cpp"
+#include "qgsbackgroundcachedfeaturesource.h"
 #include "qgsbackgroundcachedshareddata.h"
 
 #include "qgsapplication.h"
 #include "qgsfeedback.h"
+#include "qgsfeaturedownloader.h"
 #include "qgslogger.h"
 #include "qgsmessagelog.h"
 #include "qgswfsutils.h" // for isCompatibleType()
 #include "qgsgeometryengine.h"
+#include "qgsvectordataprovider.h"
 
 #include <QDataStream>
 #include <QDir>
 #include <QFile>
 #include <QJsonDocument>
 #include <QMutex>
-#include <QPushButton>
-#include <QStyle>
 #include <QTimer>
 #include <QElapsedTimer>
-#include <QThreadPool>
 
-#include <sqlite3.h>
-
-const QString QgsBackgroundCachedFeatureIteratorConstants::FIELD_GEN_COUNTER( QStringLiteral( "__qgis_gen_counter" ) );
-const QString QgsBackgroundCachedFeatureIteratorConstants::FIELD_UNIQUE_ID( QStringLiteral( "__qgis_unique_id" ) );
-const QString QgsBackgroundCachedFeatureIteratorConstants::FIELD_HEXWKB_GEOM( QStringLiteral( "__qgis_hexwkb_geom" ) );
-const QString QgsBackgroundCachedFeatureIteratorConstants::FIELD_MD5( QStringLiteral( "__qgis_md5" ) );
-
-
-//
-// QgsFeatureDownloaderProgressTask
-//
-
-QgsFeatureDownloaderProgressTask::QgsFeatureDownloaderProgressTask( const QString &description, long long totalCount )
-  : QgsTask( description, QgsTask::CanCancel | QgsTask::CancelWithoutPrompt | QgsTask::Silent )
-  , mTotalCount( totalCount )
-{
-}
-
-bool QgsFeatureDownloaderProgressTask::run()
-{
-  QgsApplication::taskManager()->threadPool()->releaseThread();
-  mNotFinishedMutex.lock();
-  if ( !mAlreadyFinished )
-  {
-    mNotFinishedWaitCondition.wait( &mNotFinishedMutex );
-  }
-  mNotFinishedMutex.unlock();
-
-  QgsApplication::taskManager()->threadPool()->reserveThread();
-  return true;
-}
-
-void QgsFeatureDownloaderProgressTask::cancel()
-{
-  emit canceled();
-
-  QgsTask::cancel();
-}
-
-void QgsFeatureDownloaderProgressTask::finalize()
-{
-  const QMutexLocker lock( &mNotFinishedMutex );
-  mAlreadyFinished = true;
-
-  mNotFinishedWaitCondition.wakeAll();
-}
-
-void QgsFeatureDownloaderProgressTask::setDownloaded( long long count )
-{
-  setProgress( static_cast<double>( count ) / static_cast<double>( mTotalCount ) * 100 );
-}
-
-
-// -------------------------
-
-QgsFeatureDownloaderImpl::QgsFeatureDownloaderImpl( QgsBackgroundCachedSharedData *shared, QgsFeatureDownloader *downloader )
-  : mSharedBase( shared )
-  , mDownloader( downloader )
-{
-  // Needed because used by a signal
-  qRegisterMetaType<QVector<QgsFeatureUniqueIdPair>>( "QVector<QgsFeatureUniqueIdPair>" );
-}
-
-QgsFeatureDownloaderImpl::~QgsFeatureDownloaderImpl()
-{
-  if ( mProgressTask )
-  {
-    mProgressTask->finalize();
-    mProgressTask = nullptr;
-  }
-}
-
-void QgsFeatureDownloaderImpl::emitFeatureReceived( QVector<QgsFeatureUniqueIdPair> features )
-{
-  emit mDownloader->featureReceived( features );
-}
-
-void QgsFeatureDownloaderImpl::emitFeatureReceived( long long featureCount )
-{
-  emit mDownloader->featureReceived( featureCount );
-}
-
-void QgsFeatureDownloaderImpl::emitEndOfDownload( bool success )
-{
-  emit mDownloader->endOfDownload( success );
-}
-
-void QgsFeatureDownloaderImpl::emitResumeMainThread()
-{
-  emit mDownloader->resumeMainThread();
-}
-
-void QgsFeatureDownloaderImpl::stop()
-{
-  QgsDebugMsgLevel( QStringLiteral( "QgsFeatureDownloaderImpl::stop()" ), 4 );
-  mStop = true;
-  emitDoStop();
-}
-
-void QgsFeatureDownloaderImpl::setStopFlag()
-{
-  QgsDebugMsgLevel( QStringLiteral( "QgsFeatureDownloaderImpl::setStopFlag()" ), 4 );
-  mStop = true;
-}
-
-
-// Called from GUI thread
-void QgsFeatureDownloaderImpl::createProgressTask( long long numberMatched )
-{
-  Q_ASSERT( qApp->thread() == QThread::currentThread() );
-
-  // Make sure that the creation is done in an atomic way, so that the
-  // starting thread (running QgsFeatureDownloaderImpl::run()) can be sure that
-  // this function has either run completely, or not at all (mStop == true),
-  // when it wants to destroy mProgressTask
-  QMutexLocker locker( &mMutexCreateProgressTask );
-
-  if ( mStop )
-    return;
-  Q_ASSERT( !mProgressTask );
-
-  mProgressTask = new QgsFeatureDownloaderProgressTask( QObject::tr( "Loading features for layer %1" ).arg( mSharedBase->layerName() ), numberMatched );
-  QgsApplication::taskManager()->addTask( mProgressTask );
-}
-
-void QgsFeatureDownloaderImpl::endOfRun( bool serializeFeatures, bool success, int totalDownloadedFeatureCount, bool truncatedResponse, bool interrupted, const QString &errorMessage )
-{
-  {
-    QMutexLocker locker( &mMutexCreateProgressTask );
-    mStop = true;
-  }
-
-  if ( serializeFeatures )
-    mSharedBase->endOfDownload( success, totalDownloadedFeatureCount, truncatedResponse, interrupted, errorMessage );
-  else if ( !errorMessage.isEmpty() )
-    mSharedBase->pushError( errorMessage );
-
-  // We must emit the signal *AFTER* the previous call to mShared->endOfDownload()
-  // to avoid issues with iterators that would start just now, wouldn't detect
-  // that the downloader has finished, would register to itself, but would never
-  // receive the endOfDownload signal. This is not just a theoretical problem.
-  // If you switch both calls, it happens quite easily in Release mode with the
-  // test suite.
-  emitEndOfDownload( success );
-
-  if ( mProgressTask )
-  {
-    mProgressTask->finalize();
-    mProgressTask = nullptr;
-  }
-
-  if ( mTimer )
-  {
-    mTimer->deleteLater();
-    mTimer = nullptr;
-  }
-}
-
-// -------------------------
-
-
-void QgsFeatureDownloader::run( bool serializeFeatures, long long maxFeatures )
-{
-  Q_ASSERT( mImpl );
-  mImpl->run( serializeFeatures, maxFeatures );
-}
-
-void QgsFeatureDownloader::stop()
-{
-  Q_ASSERT( mImpl );
-  mImpl->stop();
-}
-
-// -------------------------
-
-QgsThreadedFeatureDownloader::QgsThreadedFeatureDownloader( QgsBackgroundCachedSharedData *shared )
-  : mShared( shared )
-  , mRequestMadeFromMainThread( QThread::currentThread() == QApplication::instance()->thread() )
-{
-}
-
-QgsThreadedFeatureDownloader::~QgsThreadedFeatureDownloader()
-{
-  stop();
-}
-
-void QgsThreadedFeatureDownloader::stop()
-{
-  if ( mDownloader )
-  {
-    mDownloader->stop();
-    wait();
-    delete mDownloader;
-    mDownloader = nullptr;
-  }
-}
-
-void QgsThreadedFeatureDownloader::startAndWait()
-{
-  start();
-
-  QMutexLocker locker( &mWaitMutex );
-  while ( !mDownloader )
-  {
-    mWaitCond.wait( &mWaitMutex );
-  }
-}
-
-void QgsThreadedFeatureDownloader::run()
-{
-  // We need to construct it in the run() method (i.e. in the new thread)
-  mDownloader = new QgsFeatureDownloader();
-  mDownloader->setImpl( mShared->newFeatureDownloaderImpl( mDownloader, mRequestMadeFromMainThread ) );
-  {
-    QMutexLocker locker( &mWaitMutex );
-    mWaitCond.wakeOne();
-  }
-  mDownloader->run( true, /* serialize features */
-                    mShared->requestLimit() /* user max features */ );
-}
+const QString QgsBackgroundCachedFeatureIterator::Constants::FIELD_GEN_COUNTER( QStringLiteral( "__qgis_gen_counter" ) );
+const QString QgsBackgroundCachedFeatureIterator::Constants::FIELD_UNIQUE_ID( QStringLiteral( "__qgis_unique_id" ) );
+const QString QgsBackgroundCachedFeatureIterator::Constants::FIELD_HEXWKB_GEOM( QStringLiteral( "__qgis_hexwkb_geom" ) );
+const QString QgsBackgroundCachedFeatureIterator::Constants::FIELD_MD5( QStringLiteral( "__qgis_md5" ) );
 
 // -------------------------
 
@@ -422,7 +206,7 @@ QgsFeatureRequest QgsBackgroundCachedFeatureIterator::initRequestCache( int genC
     }
     if ( genCounter >= 0 )
     {
-      requestCache.combineFilterExpression( QString( QgsBackgroundCachedFeatureIteratorConstants::FIELD_GEN_COUNTER + " <= %1" ).arg( genCounter ) );
+      requestCache.combineFilterExpression( QString( QgsBackgroundCachedFeatureIterator::Constants::FIELD_GEN_COUNTER + " <= %1" ).arg( genCounter ) );
     }
   }
 
@@ -494,7 +278,7 @@ void QgsBackgroundCachedFeatureIterator::fillRequestCache( QgsFeatureRequest req
 
     if ( mFetchGeometry )
     {
-      int hexwkbGeomIdx = dataProviderFields.indexFromName( QgsBackgroundCachedFeatureIteratorConstants::FIELD_HEXWKB_GEOM );
+      int hexwkbGeomIdx = dataProviderFields.indexFromName( QgsBackgroundCachedFeatureIterator::Constants::FIELD_HEXWKB_GEOM );
       Q_ASSERT( hexwkbGeomIdx >= 0 );
       cacheSubSet.append( hexwkbGeomIdx );
     }
@@ -639,7 +423,7 @@ bool QgsBackgroundCachedFeatureIterator::fetchFeature( QgsFeature &f )
 
     if ( mShared->hasGeometry() && mFetchGeometry )
     {
-      int idx = cachedFeature.fields().indexFromName( QgsBackgroundCachedFeatureIteratorConstants::FIELD_HEXWKB_GEOM );
+      int idx = cachedFeature.fields().indexFromName( QgsBackgroundCachedFeatureIterator::Constants::FIELD_HEXWKB_GEOM );
       Q_ASSERT( idx >= 0 );
 
       const QVariant &v = cachedFeature.attributes().value( idx );
@@ -873,7 +657,7 @@ bool QgsBackgroundCachedFeatureIterator::rewind()
     QgsFeatureRequest requestCache;
     int genCounter = mShared->getUpdatedCounter();
     if ( genCounter >= 0 )
-      requestCache.setFilterExpression( QString( QgsBackgroundCachedFeatureIteratorConstants::FIELD_GEN_COUNTER + " <= %1" ).arg( genCounter ) );
+      requestCache.setFilterExpression( QString( QgsBackgroundCachedFeatureIterator::Constants::FIELD_GEN_COUNTER + " <= %1" ).arg( genCounter ) );
     else
       mDownloadFinished = true;
     auto cacheDataProvider = mShared->cacheDataProvider();
@@ -952,19 +736,4 @@ void QgsBackgroundCachedFeatureIterator::copyFeature( const QgsFeature &srcFeatu
   dstFeature.setValid( true );
   dstFeature.setId( srcFeature.id() );
   dstFeature.setFields( fields ); // allow name-based attribute lookups
-}
-
-
-// -------------------------
-
-QgsBackgroundCachedFeatureSource::QgsBackgroundCachedFeatureSource(
-  std::shared_ptr<QgsBackgroundCachedSharedData> shared
-)
-  : mShared( std::move( shared ) )
-{
-}
-
-QgsFeatureIterator QgsBackgroundCachedFeatureSource::getFeatures( const QgsFeatureRequest &request )
-{
-  return QgsFeatureIterator( new QgsBackgroundCachedFeatureIterator( this, false, mShared, request ) );
 }
