+++ src/providers/wfs/qgswfsprovider.cpp
@@ -18,7 +18,9 @@
 #include "qgis.h"
 #include "qgscplhttpfetchoverrider.h"
 #include "qgssetrequestinitiator_p.h"
+#include "qgsbackgroundcachedfeaturesource.h"
 #include "qgsfeature.h"
+#include "qgsfeaturedownloader.h"
 #include "qgsfeedback.h"
 #include "qgsfields.h"
 #include "qgsgeometry.h"
